requirements.txt
# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Twilio SMS
twilio==8.10.0

# HTTP requests
requests==2.31.0

# Data handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Retry functionality
tenacity==8.2.3

# Phone number validation
phonenumbers==8.13.24

# Email utilities
email-validator==2.1.0 